Debug Netlify Function: Print request and context
The issue with the compare page lies in the use of the # symbol. The #
symbol is understood as the end of the URL and so only the content up to
the hash are passed to params.

It's not clear to me whether this error happens at the Netlify level,
next-on-netlify/wrapper, or the NextJS level. Let's see if Netlify
passes the correct URL to next-on-netlify.